lab3/111511198/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
blas=1.0=mkl
bottleneck=1.3.7=py310ha9d4c09_0
brotli-python=1.0.9=py310h6a678d5_7
bzip2=1.0.8=h5eee18b_5
ca-certificates=2024.3.11=h06a4308_0
certifi=2024.2.2=py310h06a4308_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
cuda-cudart=12.1.105=0
cuda-cupti=12.1.105=0
cuda-libraries=12.1.0=0
cuda-nvrtc=12.1.105=0
cuda-nvtx=12.1.105=0
cuda-opencl=12.4.99=0
cuda-runtime=12.1.0=0
cudatoolkit=11.1.1=hb139c0e_13
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py310h06a4308_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.10.0=py310h06a4308_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
huggingface_hub=0.20.3=py310h06a4308_0
idna=3.4=py310h06a4308_0
importlib-metadata=7.0.1=py310h06a4308_0
intel-openmp=2023.1.0=hdb19cb5_46306
jinja2=3.1.3=py310h06a4308_0
jpeg=9e=h5eee18b_1
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=12.1.0.26=0
libcufft=11.0.2.4=0
libcufile=1.9.0.20=0
libcurand=10.3.5.119=0
libcusolver=11.4.4.55=0
libcusparse=12.0.2.55=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=13.2.0=h807b86a_5
libgomp=13.2.0=h807b86a_5
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=12.0.2.50=0
libnvjitlink=12.1.105=0
libnvjpeg=12.1.1.14=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=13.2.0=h7e041cc_5
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.3=py310h5eee18b_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.8=py310h5eee18b_0
mkl_random=1.2.4=py310hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py310h06a4308_0
numexpr=2.8.7=py310h85018f9_0
numpy=1.26.4=py310h5f9d8c6_0
numpy-base=1.26.4=py310hb5e798b_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.2.1=hd590300_1
packaging=23.2=py310h06a4308_0
pandas=2.2.1=py310h6a678d5_0
pillow=10.2.0=py310h5eee18b_0
pip=23.3.1=py310h06a4308_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.14=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytorch=2.2.2=py3.10_cuda12.1_cudnn8.9.2_0
pytorch-cuda=12.1=ha16c6d3_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=py310h06a4308_0
pyyaml=6.0.1=py310h5eee18b_0
readline=8.2=h5eee18b_0
regex=2023.10.3=py310h5eee18b_0
requests=2.31.0=py310h06a4308_1
safetensors=0.4.2=py310ha89cbab_0
setuptools=68.2.2=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
sympy=1.12=py310h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.15.1=py310h22610ee_0
torchaudio=2.2.2=py310_cu121
torchtriton=2.2.0=py310
torchvision=0.17.2=py310_cu121
tqdm=4.65.0=py310h2f386ee_0
transformers=4.37.2=py310h06a4308_0
typing-extensions=4.9.0=py310h06a4308_1
typing_extensions=4.9.0=py310h06a4308_1
tzdata=2024a=h04d1e81_0
urllib3=2.1.0=py310h06a4308_1
wheel=0.41.2=py310h06a4308_0
xz=5.4.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.17.0=py310h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
